+++ CHANGES.rst
@@ -1,5 +1,5 @@
 Changes
-^^^^^^^
+=======
 
 This is a record of all past pybaum releases and what went into them in reverse
 chronological order. We follow `semantic versioning <https://semver.org/>`_ and all
@@ -11,6 +11,8 @@ releases are available on `Anaconda.org
 ------------------
 
 - :gh:`2` replaces the pre-commit pipeline step with pre-commit.ci.
+- :gh:`11` implement :func:`pybaum.tree_util.tree_yield` and
+  :func:`pybaum.tree_util.tree_just_yield`.
 - :gh:`10` adds default arguments to ``tree_just_flatten``.
 - :gh:`12` adds a section about the API to the docs.